Big warnings cleanup for Solaris/GCC. Down to about 40 now, but we'll get there one day.
Use `cat' to create aclocal.m4, not `aclocal'. Some people don't
have automake installed.

Only run the autoconf rule in the top-level GNUmakefile if the
invoker specified `make configure', don't run it automatically
because of CVS timestamp skew.

+++ src/backend/commands/copy.c
@@ -325,7 +325,7 @@ DoCopy(char *relname, bool binary, bool oids, bool from, bool pipe,
 				elog(ERROR, "COPY command, running in backend with "
 					 "effective uid %d, could not open file '%s' for "
 					 "reading.  Errno = %s (%d).",
-					 geteuid(), filename, strerror(errno), errno);
+					 (int) geteuid(), filename, strerror(errno), errno);
 		}
 		CopyFrom(rel, binary, oids, fp, delim, null_print);
 	}
@@ -358,7 +358,7 @@ DoCopy(char *relname, bool binary, bool oids, bool from, bool pipe,
 				elog(ERROR, "COPY command, running in backend with "
 					 "effective uid %d, could not open file '%s' for "
 					 "writing.  Errno = %s (%d).",
-					 geteuid(), filename, strerror(errno), errno);
+					 (int) geteuid(), filename, strerror(errno), errno);
 		}
 		CopyTo(rel, binary, oids, fp, delim, null_print);
 	}
