+++ src/backend/utils/adt/varlena.c
@@ -57,9 +57,9 @@ byteain(char *inputText)
 		{
 			if (*tp == '\\')
 				tp++;
-			else if (!isdigit(*tp++) ||
-					 !isdigit(*tp++) ||
-					 !isdigit(*tp++))
+			else if (!isdigit((int) *tp++) ||
+					 !isdigit((int) *tp++) ||
+					 !isdigit((int) *tp++))
 				elog(ERROR, "Bad input string for type bytea");
 		}
 	tp = inputText;
@@ -111,7 +111,7 @@ byteaout(bytea *vlena)
 	for (i = vlena->vl_len - VARHDRSZ; i != 0; i--, vp++)
 		if (*vp == '\\')
 			len += 2;
-		else if (isascii(*vp) && isprint(*vp))
+		else if (isascii((int) *vp) && isprint((int) *vp))
 			len++;
 		else
 			len += VARHDRSZ;
@@ -124,7 +124,7 @@ byteaout(bytea *vlena)
 			*rp++ = '\\';
 			*rp++ = '\\';
 		}
-		else if (isascii(*vp) && isprint(*vp))
+		else if (isascii((int) *vp) && isprint((int) *vp))
 			*rp++ = *vp++;
 		else
 		{
