+++ src/utils/version.c
@@ -76,7 +76,7 @@ ValidatePgVersion(const char *path, char **reason_p)
 	{
 		nread = read(fd, version, sizeof(version) - 1);
 		if (nread < 4 ||
-			!isdigit(version[0]) ||
+			!isdigit((int)version[0]) ||
 			version[nread - 1] != '\n')
 		{
 			*reason_p = malloc(100 + strlen(full_path));
